debugging

+++ examples/controlnet/train_controlnet_sdxl.py
@@ -84,10 +84,10 @@ def log_validation(controlnet, args, accelerator, step):
     # https://github.com/huggingface/diffusers/pull/4038#issuecomment-1631245691
     pipeline = StableDiffusionXLControlNetPipeline.from_pretrained(
         args.pretrained_model_name_or_path,
-        controlnet=controlnet,
+        controlnet=controlnet.to(torch.float16),
         safety_checker=None,
         revision=args.revision,
-        torch_dtype=torch.float32,
+        torch_dtype=torch.float16,
     )
     pipeline.scheduler = UniPCMultistepScheduler.from_config(pipeline.scheduler.config)
     pipeline = pipeline.to(accelerator.device)
